Add -x parameter for extracting the contents of the crouton bundle.

+++ build/wrapper.sh
@@ -32,16 +32,28 @@
 
 # This is a wrapped tarball. This script untars itself to a temporary directory
 # and then runs installer/main.sh with the parameters passed to it.
+# You can pass -x [directory] to extract the contents somewhere.
 
-# Make a temporary directory and auto-remove it when the script ends.
-SCRIPTDIR="`mktemp -d --tmpdir=/tmp "${0##*/}.XXX"`"
-TRAP="rm -rf \"$SCRIPTDIR\"; $TRAP"
-trap "$TRAP" INT HUP 0
+if [ "$1" = '-x' -a "$#" -le 2 ]; then
+    # Extract to the specified directory.
+    SCRIPTDIR="${2:-"${0##*/}.unbundled"}"
+    mkdir -p "$SCRIPTDIR"
+else
+    # Make a temporary directory and auto-remove it when the script ends.
+    SCRIPTDIR="`mktemp -d --tmpdir=/tmp "${0##*/}.XXX"`"
+    TRAP="rm -rf \"$SCRIPTDIR\"; $TRAP"
+    trap "$TRAP" INT HUP 0
+fi
 
 # Extract this file after the ### line
 # TARPARAMS will be set by the Makefile to match the compression method.
 line="`awk '/^###/ { print FNR+1; exit 0; }' "$0"`"
 tail -n "+$line" "$0" | tar -x $TARPARAMS -C "$SCRIPTDIR"
+
+# Exit here if we're just extracting
+if [ -z "$TRAP" ]; then
+    exit
+fi
 
 # Execute the main script inline. It will use SCRIPTDIR to find what it needs.
 . "$SCRIPTDIR/installer/main.sh"
